pyproject.toml
[tool.poetry]
name = "sentiment-example"
version = "0.1.0"
description = ""
authors = ["jihoahn9303 <rediate03@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
scikit-learn = "^1.4.0"
matplotlib = "^3.8.2"
pandas = "^2.2.0"
numpy = "^1.26.3"
mlflow = "^2.10.0"
fastapi = "^0.109.2"
google-cloud-storage = "^2.14.0"
hydra-core = "^1.3.2"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.27"
pymysql = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
